Change theme field names and add comments
I changed the editor_fg and editor_bg fields to app_fg and app_bg because these fields dictate the color of the entire app, not just the editor space. I also added comments to the theme struct to clarify what each field controls, and I add clarifying comments to the config module.