Move after_cursor logic to own function
The logic get the rest of the text on the line after the cursor is now in its own function. This is used in the enter and backspace functions. Additionally, enter now works properly: it wouldn't move the last character to the next line if  it was the only thing to be moved.