Create backspace and fix save
The backspace key now works as expected. The save feature now works properly. It wasn't a problem before because text wasn't being deleted before backspace, but backspacing wouldn't always be saved. Now, a new blank file is created each time and the content is written to it, so it will always save correctly.