Add operation enum and put move cursor in own func
The logic for moving the cursor to a different line (for the up and down arrow functions) is now in its own function (move_cursor_line). The enum Operation is also added which contains the operations ADD and SUB to track which operation is used in move_cursor_line. The PartialEq trait is also implemented for the Operation enum to compare operations.