Implement delete_selection
The highlighted selection of text can now be deleted using the `delete_selection` function in the editor module. Different keys (insert character ,enter, backspace, etc.) now also call this function to automatically delete the selection when pressed. The `Selection` struct also now tracks the starting position in the text and the starting position of the cursor so that they can be reset after the selection is deleted. Lastly, a few tests have been implemented for the `delete_selection`.