Clean up code
All of the selection initializations have been added to two functions: `init_forward` for right, down, end and `init_backward` for left, up, home. Also, there is now a function called `end_helper` which contains the (non-init) highlight logic for end_highlight. This logic is long, so I put it in a helper function to improved readability.