Fix unicode edge case
Previously, if a TextBlock was loaded in such that it began or ended in the middle of a unicode character (because they are multiple bytes), the program would panic. This is because the str::from_utf8 checks for valid unicode characters, so now an unsafe block is used to use the str::from_utf8_unchecked function to disable this check.

- This is unrelated to undo-redo, I just wanted to fix this real quick.