Implement a runtime benchmark (Issue #32)
Added a benchmark to test the runtime of this version versus the version
of the app from commit `55730f20d18c2a37d74aabe678ff74c56b34b936`. This
version completed the benchmark in `~33 seconds` while the other version
completed it in about the same, so the tracker created a negligible
improvement. However, I'm keeping the tracker because I still prefer it
over recalculating the location regularly.